+++ hooks/utils/productImportsHelpers.ts
@@ -1,6 +1,9 @@
 import { TenantProduct } from "@/context/TenantContext/types";
-import { SetStateAction } from "react";
-import { CsvPreviewProps, JsonPreviewProps } from "../types";
+
+import {
+  _importProductsCSV,
+  _importProductsJSON,
+} from "./tenantOnboardingHelpers";
 
 /*
 FIELD CANDIDATES
@@ -146,98 +149,9 @@ const getPaginationMeta = (
   };
 };
 
-const _validateProducts = ({
-  products,
-  setErrors,
-  setIsValidated,
-}: {
-  products: TenantProduct[];
-  setErrors: (value: SetStateAction<string[]>) => void;
-  setIsValidated: (value: SetStateAction<boolean>) => void;
-}) => {
-  const newErrors: string[] = [];
-  products.forEach((p, index) => {
-    if (!p.name) newErrors.push(`Product ${index + 1} missing name`);
-    if (!p.url) newErrors.push(`Product ${index + 1} missing url`);
-    if (p.price !== null && p.price != undefined && p.price < 0)
-      newErrors.push(`Product ${index + 1} invalid price`);
-  });
-
-  if (newErrors.length > 0) {
-    setErrors(newErrors);
-    setIsValidated(false);
-    return false;
-  }
-
-  setIsValidated(true);
-  return true;
-};
-
-const _handleFileUpload = async ({
-  file,
-  setErrors,
-  setFileName,
-  setFile,
-  setJsonPreview,
-  setCsvPreview,
-  setProducts,
-  importProductsJSON,
-  importProductsCSV,
-  maxProducts,
-}: {
-  file: File | null;
-  setErrors: (value: SetStateAction<string[]>) => void;
-  setFileName: (value: SetStateAction<string | null>) => void;
-  setFile: (value: SetStateAction<File | null>) => void;
-  setProducts: (value: SetStateAction<TenantProduct[]>) => void;
-  setJsonPreview: (value: SetStateAction<JsonPreviewProps>) => void;
-  setCsvPreview: (value: SetStateAction<CsvPreviewProps>) => void;
-  importProductsJSON:
-    | ((file: File, dryRun?: boolean) => Promise<unknown>)
-    | undefined;
-  importProductsCSV:
-    | ((file: File, dryRun?: boolean) => Promise<unknown>)
-    | undefined;
-  maxProducts: 100 | 200 | 500 | 50;
-}) => {
-  if (!file) return;
-  setErrors([]);
-  setFileName(file.name);
-  setFile(file);
-
-  if (file.name.endsWith(".json") && importProductsJSON) {
-    const jPreview = (await importProductsJSON(file, true)) as {
-      total: number;
-      valid: number;
-      preview: unknown[];
-      products: unknown[];
-      errors: string[];
-    };
-    if (jPreview.products.length > maxProducts) {
-      setErrors([`Plan allows only ${maxProducts} products`]);
-      return;
-    }
-    const productsArray = jPreview.products as TenantProduct[];
-
-    setProducts(productsArray);
-    setJsonPreview(jPreview);
-  }
-  if (file.name.endsWith(".csv") && importProductsCSV) {
-    const cPreview = (await importProductsCSV(file, true)) as {
-      preview: unknown[];
-      errors: string[];
-      total: number;
-      valid: number;
-    };
-    setCsvPreview(cPreview);
-  }
-};
-
 export {
   getPaginationMeta,
   paginateProducts,
   normalizeProducts,
   PRODUCTS_PER_PAGE,
-  _validateProducts,
-  _handleFileUpload,
 };
